add AbstractVectorLayer and extend WFS and VectorTileLayer
