+++ bundles/mapping/mapwfs2/domain/WFSLayer.js
@@ -1,13 +1,11 @@
-import { VectorStyle } from '../../mapmodule/domain/VectorStyle';
+import { AbstractVectorLayer } from '../../mapmodule/domain/AbstractVectorLayer';
 /**
  * @class Oskari.mapframework.bundle.mapwfs2.domain.WFSLayer
  *
  * MapLayer of type WFS
  */
 
-const VectorTileLayer = Oskari.clazz.get('Oskari.mapframework.mapmodule.VectorTileLayer');
-
-export class WFSLayer extends VectorTileLayer {
+export class WFSLayer extends AbstractVectorLayer {
     constructor () {
         super(...arguments);
         /* Layer Type */
@@ -28,10 +26,6 @@ export class WFSLayer extends VectorTileLayer {
         return true;
     }
 
-    getLegendImage () {
-        return null;
-    }
-
     isSupportedSrs () {
         return true;
     }
@@ -41,21 +35,6 @@ export class WFSLayer extends VectorTileLayer {
     }
 
     /* Layer type specific s */
-
-    createStylesFromOptions (clear) {
-        if (clear === true) {
-            this._styles = [];
-        }
-        // Read options object for styles and hover options
-        const options = this.getOptions() || {};
-        const { styles = {} } = options;
-        Object.keys(styles).forEach(styleId => {
-            const style = new VectorStyle(styleId, null, 'normal', styles[styleId]);
-            this.addStyle(style);
-        });
-        // Remove styles from options to be sure that VectorStyle is used
-        delete options.styles;
-    }
 
     /**
      * @method getFields
